refactor(ses): Create ses/compartment entry
